# Read and cache regexes when loading the database