The queries for regular expressions that match browser, OS, device, and
fragments are not parameterized, and so can be read once and stored on
the Parser to avoid iterating a SQLite query cursor each time.

A test file containing 36404 non-unique user agents extracted from real
server logs, which took 155 seconds to parse before, now takes 73
seconds.  Significant benefits are still seen when memoizing parses in
client code; processing time decreases from 17 seconds to 10 seconds.
(This time does not include fragment parsing, which sees a similar
magnitude speed-up.)

Removing the @ silencing operator from the preg_match calls also has a
minor but measurable benefit (73 to 70 seconds), and all such calls
outside fragment parsing should only be receiving valid regular
expressions and strings to begin with.